// frontend/src/api/models/VideoProcessingApplication.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const VideoProcessingApplication = {
    GenerateShorts: 'generate_shorts',
    ApplyTrim: 'apply_trim',
    ApplySubtitles: 'apply_subtitles',
    ApplyVoiceOver: 'apply_voice_over',
    ApplyVlog: 'apply_vlog',
    ApplyPodcastTemplate: 'apply_podcast_template'
} as const;
export type VideoProcessingApplication = typeof VideoProcessingApplication[keyof typeof VideoProcessingApplication];


export function instanceOfVideoProcessingApplication(value: any): boolean {
    for (const key in VideoProcessingApplication) {
        if (Object.prototype.hasOwnProperty.call(VideoProcessingApplication, key)) {
            if (VideoProcessingApplication[key as keyof typeof VideoProcessingApplication] === value) {
                return true;
            }
        }
    }
    return false;
}

export function VideoProcessingApplicationFromJSON(json: any): VideoProcessingApplication {
    return VideoProcessingApplicationFromJSONTyped(json, false);
}

export function VideoProcessingApplicationFromJSONTyped(json: any, ignoreDiscriminator: boolean): VideoProcessingApplication {
    return json as VideoProcessingApplication;
}

export function VideoProcessingApplicationToJSON(value?: VideoProcessingApplication | null): any {
    return value as any;
}

export function VideoProcessingApplicationToJSONTyped(value: any, ignoreDiscriminator: boolean): VideoProcessingApplication {
    return value as VideoProcessingApplication;
}